Backend/requirements.txt
flask==3.0.2
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
psycopg2-binary==2.9.9
pillow==10.2.0
torch~=2.5.1
transformers==4.37.2
python-dotenv==1.0.1
werkzeug==3.0.1
numpy~=2.1.3
requests~=2.32.3
PyJWT~=2.10.1
oss2~=2.19.1
SQLAlchemy~=2.0.36